refactor: reorganization
